kriek/requirements.txt
Django==1.6.2
django-suit==0.2.5
djangorestframework==2.3.12
gunicorn==18.0
psycopg2==2.5.2
wsgiref==0.1.2
